OPSMN-4173: Fixed the specimens API to sort specimen details list based on the input query attribute (labels/barcodes/IDs).
When specimen barcodes are scanned, the UI makes an API call to obtain
the specimen details. This specimens list is sorted based on ID/label.
This ordering may not be same as the order in which barcodes are
inputted in container locations map resulting in random positions
being assigned to the specimens.

Fixed the API so that the resultant list is sorted based on input
parameter.

(Note: the specimen details are retrieved to ensure there are
no conflicting specimens having same label/barcode in different
CPs. Prior to v3.4, no such API was invoked).