OPSMN-4166: Set only one osAuthToken cookie in response header after successful sign-in.
The issue is - OS was setting 2 cookies on successful sign-in in following order:

A) clear old cookie, if any
B) new cookie to be used

It seems the older browsers ignored the 2nd cookie thereby resulting in basic
auth pop-up being displayed for all views that depend on cookies for successful
authentication with OS APIs.

Fix: Set only one token cookie in the response.

1) If unsuccessul sign-in or invalid token, clear the old cookie. (A)
2) For successful sign-in, set only the new cookie. (B)