OPSMN-4252: Stick the specimen tree table to the top of containing pane.
When users scroll the visit or any page containing specimen tree, the
view will scroll until the specimen tree reaches the top of the pane.
After which the specimen tree table sticks to the top and gets
vertical and horizontal scroll bars to allow users scroll across
the height and breadth of the tree.