OPSMN-4212: Used ng-click instead of ng-change to notify the change in checkbox status.
The intermediate tree nodes do not have the model values (true/false) set. Their
checkboxes are ticked based on whether any of their descendant nodes are ticked or not.
As a result, when ticked nodes are clicked for first time, their model value is set to
false. This will not trigger ng-change handler and as a result there will be no change
in checkbox status of its descendants.

Use of ng-click will invoke the handler everytime the checkbox is clicked irrespective
of its state.