OPSMN-3635: Fixed the conditional logic to determine whether to display specimen related buttons on results view or not.
With addition of aggregate and reporting options, each selected field is an object
containing the field name, and aggregate functions applied on the field. However,
the logic expected selected fields to be a list of simple string elements causing
the conditional logic to always evaluate to false. Hence the bug.

The fix involves inspecting whether the selected field element is simple string
or object and handle it accordingly.