OPSMN-4270: Sort/order the matching PVs based on the starting position of the PV substring that matches the input search string.
Problem:
1. The PV API uses Hibernate's criteria object to find the matching PVs.
The criteria object is used because it allows forming the search conditions
dynamically or at runtime.
2. Hibernate provides Order object to order results based on simple property
values either in asc or desc order.
3. However, we need a little bit more complex ordering mechanism where
the results are ordered based on the starting position of the substring
match.

For this purpose, implemented OrderBySubstringMatch which extends
Hibernate's Order object and generates the equivalent SQL fragment
that works on both Oracle and MySQL DB.

Syntax:
1. OrderBySubstringMatch.asc(propertyName, searchString).ignoreCase();
2. OrderBySubstringMatch.desc(propertyName, searchString).ignoreCase();