DE JAR with following fixes:
OPSMN-3594: Raise an error when more than 1K values are used for
'is one of' and 'is not one of' filters.

The problem is the underlying database like Oracle do not allow
more than 1K values for 'in' and 'not in' clauses. Besides that
they can result in performance issues.

OPSMN-2922: Fixed the FieldNode's equals() method which was causing NPE.