OPSMN-4278: Renamed the column used for storing patient UID like SSN etc to NUID.
The problem with using UID is Oracle has a built-in function with same name UID()
Therefore it gets confused as to whether the user is referring to the built-in
function or the column. To fix the ambiguity renamed the column in staging
participants table to NUID (i.e. National Unique ID)