OPSMN-4259: Added a guard to ensure null requestor is not accessed during creation of order object.
The problem was the code responsible for ensuring the receiving site belongs to the
requestor institute was accessing a null requestor object because of which NPE was
raised.  The requester object is null when the API is not able to find the valid user
from a given email ID / username.