OPSMN-4316: Fixed the specimen position auto allocation (reservation) logic.
Prior to v5.0, it was not possible to create unplanned aliquots of different types
from the same parent specimen. Therefore the code assumed all aliquots to be of
same type. However, with changes in v5.0, this assumption was no longer valid
and therefore the auto allocation code doesn't work as desired.