Added a stored procedure to query for unique constraint name on specified columns of the table and subsequently delete the constraint.
When constraint name is not specified at the time of creation, as in
case of legacy caTissue database, Oracle assigns some random names to such
constraints. However, this creates problem when we want to delete those
constraints later, as we need to know the constraint name. This can be
quite a cumbersome task if it has to be repeated everytime. To faciliate
deletion of such constraints, implemented a stored procedure:
   os_drop_unique_constraint(<table name>, <column names>)

For example:
   call os_drop_unique_constraint('catissue_specimen', 'label')

Updated 3.3 and 4.3 schema scripts to use above stored procedure
to drop unique constraints on specimen label and barcode.