OPSMN-4354: Implementation of os-tooltip widget.
There was no tooltip widget in OS. 90% of the app views were dependent
on tooltips rendered by the browser. However, the presentation of tooltip
varied across browsers. For example, Safari renders the tooltip after
lot of delay, in very small fonts making them practically unusable.

In order to unify the tooltip presentation and make them attractive
and useful, implemented a custom tooltip widget. The usage of this
widget is given below:

<some-element os-tooltip="Hardcoded tooltip string">
  <!-- something more here -->
</some-element>

<some-element os-tooltip="{{'i18n_key' | translate}}">
  <!-- something more here -->
</some-element>