OPSMN-4389: Used app URL setting value to set the cookie path.
Earlier we were using the app context path for setting the cookie path.
However, it is not guaranteed that the path used to access the app by end users
will be same as the context path. This is very true when the app is fronted
by a web server like Apache / Nginx.