OPSMN-3063: Made all interaction objects used in import of order records as serialisable.
When importing order records (multiple row records), the importer reads all records of
the input CSV file and puts them into the disk backed cache. Initially the parsed
objects are stored in-memory, and later spooled to the disk when the threshold limit is
reached. The spooling is done by serialising the objects to a binary form. For this
purpose, all the interaction objects used in the API need to be serialisable.
However, there were couple of objects that were not serialisable and therefore the
spooling failed. This gave an impression that the importer was reading only first
few records of the input file.