WashU Feedback: Made the facet values search API case insensitive.
(The bug was not caught because usually the facet values are limited
and the search is always local. However, in rare scenarios, when facets
are based on unique attributes like PPIDs, labels, the facet values run
into large number. In such cases, the search is remote, which was case
sensitive).