OPSMN-4486: Order visits by dates if available. Otherwise, default to visit identifiers.
The new APIs ensure every visit record in DB has a visit date irrespective of its
status. However, this is not true for databases that have been upgraded from
caTissue. The pending visits were created without dates. Therefore the API
to retrieve specimens of the participant crashed when sorting the visits.

Fixed the sorting of visits so that the visit IDs are used when dates are not
available.