OPSMN-4458: Do not validate "created by" when a container is being updated.
The "created by" user field is set when the container is created. Thereafter
it is never touched. In this context, it doesn't make sense for the API to
validate whether the "created by" user is valid or not. Therefore disabled
"created by" user validation when a container is being updated.