OPSMN-4582: Invalidated the auth token cookie when a SAML based login session is terminated.
The auth token cookie remained in the browser storae even when the SAML based login
session was terminated. As a result, the subsequent interaction happened using the
invalid token. On seeing the invalid token, the user was directed back to the sign-in
page. This process continued until the user explicitly cleared the cookie from the
browser storage.