OPSMN-4642: Ensured the quick search widget labels/caption are repainted whenever the locale resources are loaded and available for use.
There was a race condition between loading of locale resources and rendering
of the quick search widget. Whenever the search widget rendering finished
before loading of the locale resources, the resource keys were displayed.

To fix the issue, re-painted the search widget whenever the locale resources
loading is finished. This means the search widget will be rendered at max 2 times.