OPSMN-4705: Implemented configurable order specimens list view. A default list view is shipped with the system workflows.
Added an option in the list view to hide the empty columns.

Merged the site configured system workflows and the default system
workflows shipped with the app. When the sites customise system
workflows, they will be missing on the newly added configurations
like the order specimens list view. In such cases the merge
functionality helps.