Enhanced export records API to run synchronously and turn off email notifications.
This is very useful when the export API is used as step in other larger use case.
For example - export a list of expiring supplies. Since the job is already
running in background, there is no point in running the export in another thread.
Also, there is no reason to send export email notification and confuse the user
when the user has not invoked it explicitly.