OPSMN-4657: Raise an error when user tries to change immutable aliquot attributes like path status, anatomic site, laterality.
This error is raised only when the user is trying to update the attributes
using bulk import mechanism.

The change is not applied for all scenarios because the UI assumes the aliquot
attributes are automatically updated by the API based on its parent value.
Fixing it for all the scenarios will definitely break the UI, which is
very costly affair to fix.