OPSMN-4426: Displayed the cards on the home page and options in the navigation menu in lexicographical order of their titles.
The home page and navigation menu are displayed using the
metadata registered by the respective modules. Earlier these
were static HTML views, which meant the order of cards and
items was pretty much fixed at the time of development.
Going forward, this change allows the cards to be arranged in
whichever fashion the user likes.

The navigation menu displays the plugin options as well.