OPSMN-4796: Fixed the timepicker CSS to ensure it doesn't exceed its parent container width limits and crash into the adjacent fields of the table layout.
The timepicker is rendered using the <table> structure. The table within
table causes the non-alignment of the field w.r.t other fields of the row.
Further, each input element of the timepicker was given a fixed width
of 50px. This meant the timepicker overflowed whenever its parent
container width was less than 100px.