OPSMN-4747: Fixed delete container API.
The container delete API was crashing with NPE when attempting to delete
the custom fields record. There was no check to determine whether the
container has any custom fields or not. In absence of custom fields,
the API crashed with NPE.

Added a conditional to delete custom fields only when they are present.