OPSMN-4803: Updated DE library with following fix in the query API.
Fixed the method to compute the common ancestor of 2 join trees.
The fix is to ensure the method doesn't traverse past the join tree
root.

Prior to the fix, the method was trying to access the parent of the
join tree's root, which is obviously not present. This was resulting
in queries to crash with NPE.