OPSMN-4898: Oracle didn't like the presence of semicolon (;) character at the end of the query when presented through the JDBC interface. It complained with an invalid character error.
Removed the extraneous semicolon (;) from the query.

Note: There are 2 queries - one for admin users who have access to all
records and other for non-admin users limiting access to only subset
of the records. The second query was problematic. Therefore the error
was noticeable only to the non-admin users.

Like all great softwares, Oracle doesn't let the users know which
is the invalid character that is causing trouble to it. As a result,
time is spent, hair are pulled apart, in finding invalid character
among all the valid ones.