OPSMN-5091: Sort/order forms in descending order of their update or creation time.
The above problem sounds a lot simpler than it is...

Problem:

1. The forms API was refactored to use Hibernate's criteria API to retrieve
and sort the matching forms. The criteria API was used because it allows
filtering forms dynamically or at the runtime. For example - list forms
used as specimen custom fields, list forms used in Lung cancer for visit
custom fields etc.

2. Hibernate criteria API's Order object allows sorting of results based
on simple property values either in ascending or descending order.

3. However, the form API needed a little bit more complex sorting mechansim
wherein the forms are sorted based on update time, if available, else sorted
on creation time. A discerning reader or SQL specialist knows that this is
not a simple order by clause. This requires conditional order by clause
like in below example:

   order by
     case
       when update_time is not null then
         update_time
       else
         creation_time
     end
   desc

Solution:

4. The above is not supported by Hibernate criteria API out of the box.
To solve the problem, implemented custom OrderByNotNullProperty, which
generates "order by" clause as in above example for any number of
input properties/attributes. The generated fragment works on any database
supported by Hibernate.

Syntax:
   OrderByNotNullProperty.desc("updateTime", "creationTime")
   OrderByNotNullProperty.desc("collectionTime", "visitTime", "regTime");

5. Updated the forms API to use OrderByNotNullProperty to sort forms
based on update or creation time.