User sign-up fixes.
The create user, and sign-up views both use the same view controller.
Recently, the create user view was updated to allow users select the
time zone from the list of available time zones. The time zones list
is retrieved using an API call, which requires a valid auth token.
However, the auth token is not available for the users doing sign-up,
which results in 401 error navigating the users to the sign-in view.

Modified the user add/edit view controller to avoid API calls like
"get TZ list" when signed-in user context is not available.