OPSMN-5112: Updated DE library with the following query generator fix:
Fixed the query generator to generate the correct joins for multi-valued
lookup fields whose parent key is not the same as the primary key of the
parent form.

Added an option to allow users specify the parent key of the multi-valued
lookup fields.

---

Updated CPR query form to use the participant ID as the parent key for
the multi-valued race and ethnicity fields. Earlier the reg ID was
used to join in the race/ethnicity tables, which was giving absurd results.