OPSMN-5128: Exposed the config setting, allow HTML mark up, to unsigned users ...
The problem was -

a) The catalog uses DE form to render the request form.
b) The DE form renderer widget queries the API server for the value of
   config setting - 'HTML mark up'.
c) Since the config setting was secured, the API returned 401 error,
   redirecting the users to the sign-in page.

Fix:

Exposed the config setting, HTML mark up, to unsigned users as well.