OPSMN-5143: Flushed the session SQLs before fetching the object form record IDs.
The problem was the inserts were cached in the session. As a result, the selects
were not able to see the cached inserts thereby resultin in incorrect evaluation
of the edit check rules. This was happening only on Oracle for which sequences
are used to auto-generate the identifiers instead of querying the tables.