OPSMN-5170: Added specimen ID column to the collection and received event tables.
Added an unique constraint on the specimen ID column of the collection and received
event tables. This ensures every primary specimen has only one collection &
received event.

The specimen ID column is initialised via a trigger that is executed when a
new DE form record entry is inserted or record entry is soft-deleted.

Added a migration script to initialise the specimen ID of the existing
collection and received events.

Works for both MySQL and Oracle.

Foundation stone for fixing the other performance related bugs.