OPSMN-5191: Fixed the pagination/chunking of specimens when distributing from the DP reserved list or cart.
Problem: The order API distributes specimens in chunks of 100 so that the
memory footprint is under control irrespective of the number of specimens
to distribute.

When the specimens are distributed, their status is updated in the database.
This means the distributed specimens don't appear in the original list or
query result.  As a result, the 2nd chunk of the original list appears as
the 1st chunk after first round of distribution. Similarly the 3rd chunk of
the original list appears as the 1st chunk after two rounds of distribution.
That is to say, the chunk to distribute is always the first one or at the top.
However, the API was incorrectly incrementing the chunk index to query and
therefore missing to distribute specimens of intermediate chunks.

The API distributed specimens of 1, 3, 6, 10 chunks and like that.

Resolution: Always query the first chunk of the list to distribute the
specimens.