OPSMN-5146: Implemented HTTP POST API to search specimens based on labels, barcodes, CP ID, PPID etc.
The newly implemented POST API is used by the specimens search UI
widget for querying the specimens based on the labels/barcodes
entered by the users. Imposed a max limit of 1000 specimen labels/barcodes
that can be searched at once.

Earlier the widget used GET API. Different browsers imposed different
limits on the URL length. We used 8K as the URL limit (LCD) and ensured
the summation of all labels length didn't exceed 8K. This meant there was
no fixed limit on the number of specimens that could be searched.
Hence the message was also vague - "too many specimens".

With the new API, the limit is definite and therefore the error message
is also definite.