OPSMN-5227: Used PV attribute as the cache key to store and retrieve the field's default list.
The default list populated in the PV fields is cached at the browser side to prevent
frequent API calls. However, this cached list was incorrectly used for PV fields of
different attributes i.e. the same default list was used for both specimen_type and
clinical_diagnosis.

To fix the problem, used PV attribute as the key to index into the cache for storing
and retrieving the field's default list.