Star feature: Starred CPs: NULL values are ordered differently by Oracle and MySQL
Oracle treats NULL as having max values than any other values appearing in the same column.
MySQL treats NULL as having min values than the other values of the column.

To ensure NULL values appear at the bottom of the list, different sort direction
needs to be used. For Oracle, it will be asc and for MySQL, it will be desc.