E-consents: Consent collector role.
a) Can register/update patient details.
b) Collect patient consents.

Other changes:

a) The carts card is displayed only if the user has read specimen rights.

b) The 'proceed to collection' button in the register participant form is
   visible only if the user has read specimen rights.

c) The "View Catalog" on each row of the CP list is displayed only if the
   user has query read rights.

d) Pre-pending specimens are created at the time of patient registration or
   visit completion without access checks.