OPSMN-5271: Implementation of rules based display of forms for the data entry.
The forms in the participant, visit, and specimen forms tab are displayed based
on the rules configured in the respective CP workflow.

When a rule matches, the forms specified in the rule are displayed. If 2 or
more rules match then the first matching rule is used. The rest of the matches
are ignored. Therefore the order of rules specified in the CP workflow matters.

If none of the rules match then all forms are displayed.

"formDataEntryRules": {
  "participant": [
    {
      "when": "user.hasRole('Consent Collector')",
      "forms": ["consent1", "consent2", "consent3"]
    },
    {
      "when": "cpr.participant.gender == 'Male'",
      "forms": ["bFormName1", "bFormName2", "bFormName3"]
    },
    {
      "when": "cpr.participant.gender == 'Female'",
      "forms": ["gFormName1", "gFormName2", "gFormName3", "gFormName4"]
    }
  ],

  "visit": [
    /* Rules for visits form data entry */
  ],

  "specimen": [
    /* Rules for specimens form data entry */
  ],

  "specimenEvent": [
    /* Rules for specimens event data entry */
  ]
}