SUPPORT-4646, SUPPORT-4486: Reinitialise the MRN search keywords.
The Oracle script to init the MRN search keywords populated the participant ID
instead of the registration ID. As a result, quick search and click based on MRN
was either navigating users to the incorrect record or resulting in some UI errors.

To fix the problem, added a Oracle specific script to delete the existing
MRN search keywords and add the MRNs pointing to the correct CPR records.