OPSMN-5292: Removed the race condition between retrieving the form definition and loading latest records.
When the users select event and go to the next step, 2 API calls are
invoked simultaneously - one to obtain the event form definition and another
to load the latest records for the given (specimen, event) combination. If
the records are loaded before the form definition, the table/form was not
rendered in the last step. Majority of the times, the form definition is
retrieved before loading the records. However, there are sometimes when
the opposite happens. As a result, reproducing this issue consistently
was difficult.

Now added a barrier to ensure both the API calls are finished before
rendering the table or form in the last step.