OPSMN-5304, OPSMN-5273: Updated DE library with the following fixes:
---

OPSMN-5304: Count the deleted fields as well when generating the unique field names.
This ensures CSD UI always generates unused unique names.

Bug: The CSD UI was only counting the existing fields of the form when generating
unique names.

For example - if a form had 4 fields ST1, ST2, ST3, and ST4. Later some user
deleted 2 fields. After a while when the user tries to add a new field, CSD
generates the field name as ST3, which is either present in the existing list
or in the deleted list. This is not allowed. The users are stuck...

---

OPSMN-5273: Added foreign key constraints on the DE lookup fields/columns.
This ensures the row from the parent table cannot be deleted if it is
referenced by one or more rows of a DE table. This change will be available
only for the newly added forms and/or fields.

Examples of lookup fields include user, site, container, and PV fancy controls.