OPSMN-5325: Updated DE JAR with the following fix:
---

OPSMN-5325: Removed the restriction of max 1000 literals in the IN clause.
The restriction has been relaxed for MySQL. Removal of this restriction
doesn't allow users to include any number of literals in the IN clause. The
max will be dictated by the underlying database query parse buffer,
efficiency of database to handle large values.