OPSMN-5321: Fixed the visits and specimens builders/factories to avoid preparation of custom fields record when no collection protocol information is available.
The problem is when no PPID is specified, the factory has no idea about
the CP for which the visit is being created. However, in its attempt to
report all errors, it tries to prepare the custom fields and crashes.
Therefore the fix is to avoid custom fields preparation when no CP
information is available. This will allow the factory to report the other
accumulated errors.