OPSMN-5234: Fixed the malformed HTML in import jobs list view.
This was not caught during dev testing because testing was done
without minification and uglification (resource and time intensive op).
The HTML was served to browsers AS IS, which handled the malformed HTML
gracefully. However, during minification and uglification, the build
script tried to fix the HTML in its own way, eventually breaking the view
semantics.