OPSMN-5389: Allowed editing of shipment details in all states (Draft/Shipped/Received)
The following attributes are immutable:

a) Sending and receiving sites cannot be modified after the shipment is shipped.

b) Specimens cannot be added or removed from the shipment after it is shipped.

c) Specimen location and quantity cannot be modified after the shipment is received.

Rest of the details like shipment name, tracking URL, courier name, sender/receiver
comments, shipped and received dates, received quality can all be mutated.

Removed validation on the shipment date. It can be any date with 1/1/1970 as the baseline.