OPSMN-5397: Order UX improvements.
1. Allowed users to select upto 10K specimens from the query results view
and distribute. This was broken because the API failed to parse the
long list of specimen IDs. The API URL was rejected by the server
considering it to be very long. Used POST method to submit the IDs
and retrieve the specimen details.

2. Added a new option 'My Default Cart' to the 'Add to Cart' widget.
Clicking on this option navigates users to their default cart.

3. Displayed the error messages that occur on clicking of 'Save Draft'
or 'Distribute' in a popup or modal dialog. This allows users to
read and digest the error messages at their own pace. Further, the
popup allows users to copy the message to clipboard, which they
can use to create the support tickets with us or resolve the issues
on their own by looking at the relevant specimens appearing in the
error message.

4. Added an option in the order add/edit page to allow users remove
the specimens from the cart.