OPSMN-5359, OPSMN-5358: Updated DE library with the following fixes.
OPSMN-5359: The order of columns in the results view was distorted by
the presence of internal/hidden variables/columns created by the query
UI. The fix is to ignore such hidden columns when sorting the final
list of columns in the results view.

OPSMN-5358: Suppress the rendering of columns corresponding to the
internal variables created by the query UI. These internal variables
are used for adding or performing the actions of the correct
specimen records selected in the results view.

The fix is to send the internal variable label AS IS in the query
results so that the UI can make decision on whether to display or not.