Added an option to export the CP definition along with the IDs. The IDs are required when working with the CP in offline mode.
Earlier, this API was used to export the CP definition and recreate
it on another VM/DB. For this purpose, the IDs were not included in
the exported JSON.

This API is now used to cache the CP definition on the mobile device.
The cached CP is used in the offline mode. For this purpose, IDs are
needed for tying the specimens to their requirements etc.