1. Implemented a database driven global lock to ensure a single import job is running across all nodes of the cluster.
2. Ensured an import job can be run by only one node of the cluster.

3. Ensured a stop import job command received by any node in the cluster
is propagated to the node running the job. The max. time taken to
stop the job will be the time taken to process 25 records of the
input CSV file.

Known issue: When the node holding the global lock dies, the lock needs
to be cleared/released manually. Otherwise, no import job will run.