OPSMN-5436: Use the IP address specified in the HTTP header X-FORWARDED-FOR.
Many a times, OpenSpecimen is fronted by a proxy server like Apache, F5 etc.
The connection established with OpenSpecimen is from the proxy server. Therefore
using the connection remote address will always yield the IP address of
the proxy server. This is of no use in auditing and print rules that
are based on IP range.

To solve the above problem, almost all well behaved proxy servers capture
the IP address of the request source in the HTTP header X-FORWARDED-FOR.
The header contains the IP addresses of the nodes through which the
request passed before hitting the destination.

Added a utility method in OpenSpecimen to parse the HTTP header
X-FORWARDED-FOR, and use the first IP address as the request source.
If the address is not present then use the connection remote address
as the request source.