Updated DE JAR with the following fix
---

OPSMN-5347: Blindly setting the scale of the numbers to 2 digits after
decimal point is problematic for integers. The APIs that depend on
query APIs will break if they get 116.00 instead of just 116.

To ensure we don't set the scale of integers, added a helper routine
to check whether the number received from the database is integer
or real number. If real number then the scale is set to 2. Otherwise
the scale is set to 0.