OPSMN-5448: Updated the staged participants matching/lookup API to use database specific lookup methods.
On Oracle, the search is case insensitive. This is because Oracle
allows addition of the function based indices on the key columns when
needed.

On MySQL, no case insensitive search is employed, as the default search
done by the database is case insensitive. Employing app level case
insensitive search methods renders the indices on the key columns
useless.