Sourcing participants from external sources.
1. Added a config setting to specify one or more sites for sourcing
participants from the external system.

2. When the MRN site of the participant to be added is one of the
sourcing site then the matching participant should be present in
the staging table. Otherwise an error is raised.

3. When the existing participant MRNs are edited and MRN site is one of
the sourcing site then a matching participant should be present
in the staging table. Otherwise the edit is not permitted.

4. When the external participant source config setting is blank
or not specified, no checks are done.

5. This change is applicable both in online mode and in offline
mode done via bulk imports.